+++ examples/dialogue_bot/src/dialogue/states/start.rs
@@ -5,7 +5,7 @@ use teloxide_macros::teloxide;
 #[derive(Default)]
 pub struct StartState;
 
-#[teloxide(transition)]
+#[teloxide(subtransition)]
 async fn start(_state: StartState, cx: TransitionIn, _ans: String) -> TransitionOut<Dialogue> {
     cx.answer_str("Let's start! What's your full name?").await?;
     next(ReceiveFullNameState)